Update package version and add Clips to Channel endpoint
Increased the version number in package.json to "1.0.1-20". The main code updates revolve around the addition of a function to get clips from a channel. New interface "Clips" was created to handle this new data. The Clips function is added to the Channel endpoint which fetches clip data from the API endpoint, and a ClipsInstance is created to handle this data.